RAG/llamaindex_and_openAI/requirements.txt
llama-index
openai
pypdf
python-dotenv
pymongo
cryptography
bcrypt